requirements.txt
Flask
nest_asyncio
pandas
requests
python-telegram-bot==13.15
ta